+++ highwave/src/com/dudko/highwave/deposit/deposits/EarlierMore2Deposit.java
@@ -21,8 +21,8 @@ public class EarlierMore2Deposit extends Deposit {
 	}
 
 	@Override
-	public DepositAccount calculateDeposit(float amount, int period) {
-		float minOpenAmount = 1000000.0f;
+	public DepositAccount calculateDeposit(long amount, int period) {
+		long minOpenAmount = 1000000;
 		if (amount < minOpenAmount) {
 			return null;
 		}
