+++ highwave/src/com/dudko/highwave/deposit/deposits/SavingDeposit.java
@@ -17,7 +17,7 @@ public class SavingDeposit extends Deposit {
 	}
 
 	@Override
-	public DepositAccount calculateDeposit(float amount, int period) {
+	public DepositAccount calculateDeposit(long amount, int period) {
 		LocalDate currentDate = MinskLocalDate.now();
 		LocalDate maxEndDate = currentDate.plusYears(3);
